requirements.txt
discord==2.3.2
pyyaml==6.0.1
